根据userId查找所有的角色

+++ fo-up/src/main/java/com/fo/up/service/UpRoleService.java
@@ -16,5 +16,5 @@ public interface UpRoleService {
      * @param userId
      * @return
      */
-    public List<UpRole> findRoleByUserId(Long userId);
+    public List<UpRole> findRoleByUserId(Integer userId);
 }
